Update huston
Fixed the default color if flag_G is not provided to allow the -G flag to set the color without the default color (white) overwriting it.

I added a -z $flag_G to verify if it is empty. that way if not it wont overwrite the text color to white.